refactor: remove unused services and models from reports module
- Deleted EventTrackService and related repository and model.
- Removed ExpenseService and its repository, along with Expense model.
- Eliminated InvestmentsService, RevenueService, and their respective models and repositories.
- Cleaned up interfaces related to reports, including IAggregatedRepository and IReportsRepository.
- Removed handler and router files that were not in use.
- Updated Docker Compose files to improve service health checks and environment variable management.
- Removed unused front-end page component.

+++ api/internal/auth/model.go
@@ -5,7 +5,7 @@ import (
 )
 
 type User struct {
-	ID        int64
+	ID        uint
 	Username  string
 	Name      string
 	LastName  string
@@ -30,8 +30,8 @@ func (u User) Validate() bool {
 }
 
 type ResetToken struct {
-	ID        int64
-	UserID    int64
+	ID        uint
+	UserID    uint
 	TokenHash string
 	ExpiresAt time.Time
 }
