+++ api/pkg/DTOs/expense_dto.go
@@ -7,8 +7,8 @@ import (
 )
 
 type ExpenseDto struct {
-	ID            int64                   `json:"id"`
-	UserId        int64                   `json:"userId"`
+	ID            uint                    `json:"id"`
+	UserId        uint                    `json:"userId"`
 	Description   string                  `json:"description"`
 	Target        string                  `json:"target"`
 	Category      helpers.ExpenseCategory `json:"category"`
