+++ api/pkg/DTOs/profile.go
@@ -1,7 +1,7 @@
 package dtos
 
 type ProfileDto struct {
-	UserID    int64  `json:"user_id"`
+	UserID    uint   `json:"user_id"`
 	Email     string `json:"email"`
 	FirstName string `json:"first_name"`
 	LastName  string `json:"last_name"`
